uses/requirements.txt
aniso8601==9.0.1
apturl==0.5.2
attrs==20.3.0
awscli==1.19.1
bcrypt==3.1.7
blinker==1.4
botocore==1.20.0
Brlapi==0.8.2
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cors==1.0.1
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.5.0
distro-info==1.0
docker==4.4.4
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.20
elastic-transport==8.1.2
elasticsearch==8.2.0
fasteners==0.14.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
networkx==2.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psycopg2==2.9.3
pyasn1==0.4.8
pycairo==1.16.2
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.20.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
requests-file==1.5.1
roman==2.0.0
rsa==4.0
s3transfer==0.3.4
SecretStorage==3.3.1
selinux==3.1
simplejson==3.17.2
six==1.16.0
systemd-python==234
texttable==1.6.3
tldextract==3.3.1
urllib3==1.26.5
wadllib==1.3.5
websocket-client==0.57.0
Werkzeug==2.1.2
xdg==5
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
